accounts

+++ accounts/views.py
@@ -1,12 +1,13 @@
-from rest_framework import permissions, status, generics
+from rest_framework import permissions, status, generics, serializers
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from rest_framework_simplejwt.tokens import RefreshToken
-from .serializers import LoginSerializer, RegisterSerializer, ProfileUpdateSerializer, ProfileSerializer, ConnectionsSerializer
-from .models import Profile, Connections, CustomUser
+from rest_framework_simplejwt.tokens import RefreshToken, AccessToken
+from django.db import IntegrityError, DatabaseError
 from django.db.models import Q
-from rest_framework import serializers
-from rest_framework_simplejwt.tokens import AccessToken
+from .serializers import (
+    LoginSerializer, RegisterSerializer, ProfileUpdateSerializer, ProfileSerializer, ConnectionsSerializer
+)
+from .models import Profile, Connections, CustomUser
 
 
 class VerifyTokenView(APIView):
@@ -15,17 +16,18 @@ class VerifyTokenView(APIView):
 
     def post(self, request, *args, **kwargs):
         token = request.data.get('token')
+        if not token:
+            return Response({"detail": "Token is required."}, status=status.HTTP_400_BAD_REQUEST)
+
         try:
             access_token = AccessToken(token)
             user_id = access_token['user_id']
             user = CustomUser.objects.get(id=user_id)
-            return Response({
-                "id": user.id,
-                "username": user.username,
-                "email": user.email
-            }, status=status.HTTP_200_OK)
-        except Exception:
-            return Response({"detail": "Invalid token"}, status=status.HTTP_401_UNAUTHORIZED)
+            return Response({"id": user.id, "username": user.username, "email": user.email}, status=status.HTTP_200_OK)
+        except AccessToken.Error:
+            return Response({"detail": "Token is invalid or expired."}, status=status.HTTP_401_UNAUTHORIZED)
+        except CustomUser.DoesNotExist:
+            return Response({"detail": "User not found."}, status=status.HTTP_404_NOT_FOUND)
 
 
 class LoginView(APIView):
@@ -34,34 +36,28 @@ class LoginView(APIView):
     permission_classes = [permissions.AllowAny]
 
     def post(self, request, *args, **kwargs):
-        """Обработка POST запроса для авторизации"""
         serializer = self.serializer_class(data=request.data)
-        if serializer.is_valid():
-            user = serializer.validated_data['user']
-            refresh = RefreshToken.for_user(user)
-            access = refresh.access_token
-
-            return Response({
-                'refresh': str(refresh),
-                'access': str(access),
-            }, status=status.HTTP_200_OK)
-
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        serializer.is_valid(raise_exception=True)
+        user = serializer.validated_data['user']
+        refresh = RefreshToken.for_user(user)
+        
+        return Response({
+            'refresh': str(refresh),
+            'access': str(refresh.access_token),
+        }, status=status.HTTP_200_OK)
 
 
 class LogoutView(APIView):
     """Вью для выхода пользователя"""
-    permission_classes = (permissions.IsAuthenticated,)
+    permission_classes = [permissions.IsAuthenticated]
 
     def post(self, request, *args, **kwargs):
-        """Обработка POST запроса для выхода"""
-        try:
-            refresh_token = request.data.get("refresh_token")
-            if refresh_token is None:
-                return Response({"detail": "Refresh token is required."}, status=status.HTTP_400_BAD_REQUEST)
+        refresh_token = request.data.get("refresh_token")
+        if not refresh_token:
+            return Response({"detail": "Refresh token is required."}, status=status.HTTP_400_BAD_REQUEST)
 
-            token = RefreshToken(refresh_token)
-            token.blacklist()
+        try:
+            RefreshToken(refresh_token).blacklist()
             return Response(status=status.HTTP_205_RESET_CONTENT)
         except Exception as e:
             return Response({"detail": str(e)}, status=status.HTTP_400_BAD_REQUEST)
@@ -73,8 +69,10 @@ class RegisterView(generics.CreateAPIView):
     permission_classes = [permissions.AllowAny]
 
     def perform_create(self, serializer):
-        """Создание пользователя"""
-        serializer.save()
+        try:
+            serializer.save()
+        except IntegrityError:
+            return Response({"detail": "User with this email or username already exists."}, status=status.HTTP_400_BAD_REQUEST)
 
     def create(self, request, *args, **kwargs):
         serializer = self.get_serializer(data=request.data)
@@ -82,10 +80,7 @@ class RegisterView(generics.CreateAPIView):
         user = serializer.save()
 
         refresh = RefreshToken.for_user(user)
-        tokens = {
-            'refresh': str(refresh),
-            'access': str(refresh.access_token)
-        }
+        tokens = {'refresh': str(refresh), 'access': str(refresh.access_token)}
 
         return Response({
             'user': RegisterSerializer(user).data,
@@ -98,11 +93,11 @@ class DeleteView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def delete(self, request, *args, **kwargs):
-        """Удаление пользователя"""
-        user = request.user
-        user.delete()
-
-        return Response({"message": "User deleted successfully."}, status=status.HTTP_204_NO_CONTENT)
+        try:
+            request.user.delete()
+            return Response({"message": "User deleted successfully."}, status=status.HTTP_204_NO_CONTENT)
+        except DatabaseError:
+            return Response({"detail": "An error occurred while deleting the user."}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
 class ProfileUpdateView(generics.RetrieveUpdateAPIView):
@@ -111,26 +106,23 @@ class ProfileUpdateView(generics.RetrieveUpdateAPIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def get_object(self):
-        """Получаем профиль текущего пользователя или создаем новый"""
-        user = self.request.user
-        if not hasattr(user, 'profile'):
-            Profile.objects.create(user=user)
-        return user.profile
-    
+        try:
+            return self.request.user.profile
+        except Profile.DoesNotExist:
+            return Response({"detail": "Profile not found."}, status=status.HTTP_404_NOT_FOUND)
+
 
 class ProfileDetailView(generics.RetrieveAPIView):
+    """Представление для получения профиля пользователя"""
     serializer_class = ProfileSerializer
     permission_classes = [permissions.IsAuthenticated]
 
     def get_object(self):
-        user_id = self.kwargs.get('pk', None)
-        if user_id:
-            try:
-                user = CustomUser.objects.get(id=user_id)
-            except CustomUser.DoesNotExist:
-                raise serializers.ValidationError({"detail": "User not found"})
-        else:
-            user = self.request.user
+        user_id = self.kwargs.get('pk')
+        user = self.request.user if user_id is None else CustomUser.objects.filter(id=user_id).first()
+
+        if not user:
+            raise serializers.ValidationError({"detail": "User not found"})
 
         profile = user.profile
         profile.update_online_status()
@@ -142,53 +134,53 @@ class ContactManagementView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def post(self, request, *args, **kwargs):
-        """Отправить запрос на добавление в контакты"""
-        from_user = request.user
         to_user_id = request.data.get('to_user_id')
+        if not to_user_id:
+            return Response({"detail": "User ID is required."}, status=status.HTTP_400_BAD_REQUEST)
 
         try:
             to_user = CustomUser.objects.get(id=to_user_id)
+            if Connections.objects.filter(from_user=request.user, to_user=to_user).exists():
+                return Response({"detail": "Request already sent"}, status=status.HTTP_400_BAD_REQUEST)
+
+            connection = Connections.objects.create(from_user=request.user, to_user=to_user)
+            return Response(ConnectionsSerializer(connection).data, status=status.HTTP_201_CREATED)
         except CustomUser.DoesNotExist:
             return Response({"detail": "User not found"}, status=status.HTTP_404_NOT_FOUND)
-
-        if Connections.objects.filter(from_user=from_user, to_user=to_user).exists():
-            return Response({"detail": "Request already sent"}, status=status.HTTP_400_BAD_REQUEST)
-
-        connection = Connections.objects.create(from_user=from_user, to_user=to_user)
-        return Response(ConnectionsSerializer(connection).data, status=status.HTTP_201_CREATED)
+        except IntegrityError:
+            return Response({"detail": "Error creating connection request."}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
     def patch(self, request, *args, **kwargs):
-        """Подтвердить запрос на добавление в контакты"""
         connection_id = kwargs.get('pk')
+        if not connection_id:
+            return Response({"detail": "Connection ID is required."}, status=status.HTTP_400_BAD_REQUEST)
+
         try:
             connection = Connections.objects.get(id=connection_id, to_user=request.user, is_confirmed=False)
+            connection.is_confirmed = True
+            connection.save()
+            return Response(ConnectionsSerializer(connection).data, status=status.HTTP_200_OK)
         except Connections.DoesNotExist:
             return Response({"detail": "Request not found or already confirmed"}, status=status.HTTP_404_NOT_FOUND)
 
-        connection.is_confirmed = True
-        connection.save()
-        return Response(ConnectionsSerializer(connection).data, status=status.HTTP_200_OK)
-
     def delete(self, request, *args, **kwargs):
-        """Удалить контакт или отклонить запрос"""
         connection_id = kwargs.get('pk')
+        if not connection_id:
+            return Response({"detail": "Connection ID is required."}, status=status.HTTP_400_BAD_REQUEST)
+
         try:
             connection = Connections.objects.get(id=connection_id)
             if request.user != connection.to_user and request.user != connection.from_user:
                 return Response({"detail": "Not allowed"}, status=status.HTTP_403_FORBIDDEN)
+            connection.delete()
+            return Response({"detail": "Connection removed"}, status=status.HTTP_204_NO_CONTENT)
         except Connections.DoesNotExist:
             return Response({"detail": "Connection not found"}, status=status.HTTP_404_NOT_FOUND)
 
-        connection.delete()
-        return Response({"detail": "Connection removed"}, status=status.HTTP_204_NO_CONTENT)
-
     def get(self, request, *args, **kwargs):
-        """Получить список всех подтвержденных контактов"""
-        user = request.user
         confirmed_connections = Connections.objects.filter(
-            Q(from_user=user) | Q(to_user=user),
+            Q(from_user=request.user) | Q(to_user=request.user),
             is_confirmed=True
         )
-
         serializer = ConnectionsSerializer(confirmed_connections, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
